requirements.txt
Flask>=3.1.1
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-WTF>=1.2.1
numpy>=2.3.1
openpyxl>=3.1.5
pandas>=2.3.1
python-dotenv>=1.1.1
playwright>=1.48.0
gunicorn>=21.2.0
flask-migrate>=4.0.7
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
Flask-Compress>=1.14
Flask-Caching>=2.1.0
Flask-Limiter>=3.5.0
